requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.46.0
datasets>=2.16.0
accelerate>=0.25.0
tokenizers>=0.15.0

# Training optimization
deepspeed>=0.12.0
bitsandbytes>=0.41.0
peft>=0.7.0
flash-attn>=2.3.0

# Data processing
sentencepiece>=0.1.99
pyarrow>=14.0.0
pandas>=2.0.0
numpy>=1.24.0

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
pyyaml>=6.0
scipy>=1.11.0
scikit-learn>=1.3.0

# NVIDIA specific (optional but recommended for Nemotron)
# nemo-toolkit[all]>=1.22.0  # Uncomment if using NeMo Curator for data processing
